pull the rest of the static js into the bundle and get some tests setup
