+++ lib/client/browser-settings.js
@@ -1,8 +1,7 @@
 'use strict';
 
-var browserUtils = require('./browser-utils')();
 
-function init (serverSettings, plugins, $) {
+function init (client, plugins, serverSettings, $) {
 
   var storage = $.localStorage;
   var settings = require('../settings')();
@@ -76,7 +75,7 @@ function init (serverSettings, plugins, $) {
         storage.remove(name);
       });
       event.preventDefault();
-      browserUtils.reload();
+      client.browserUtils.reload();
     });
 
     $('#save').click(function(event) {
@@ -115,7 +114,7 @@ function init (serverSettings, plugins, $) {
       });
 
       event.preventDefault();
-      browserUtils.reload();
+      client.browserUtils.reload();
     });
   }
