+++ lib/client/careportal.js
@@ -1,20 +1,21 @@
 'use strict';
 
-(function () {
-  var translate = Nightscout.language.translate;
+var moment = require('moment-timezone');
 
-  var browserSettings = Nightscout.client.browserSettings;
-  var browserStorage = Nightscout.client.browserStorage;
+function init (client, $) {
+  var translate = client.translate;
+
+  var storage = $.localStorage;
 
   function initTreatmentDrawer() {
     $('#eventType').val('BG Check');
-    $('#glucoseValue').val('').attr('placeholder', translate('Value in') + ' ' + browserSettings.units);
+    $('#glucoseValue').val('').attr('placeholder', translate('Value in') + ' ' + client.settings.units);
     $('#meter').prop('checked', true);
     $('#carbsGiven').val('');
     $('#insulinGiven').val('');
     $('#preBolus').val(0);
     $('#notes').val('');
-    $('#enteredBy').val(browserStorage.get('enteredBy') || '');
+    $('#enteredBy').val(storage.get('enteredBy') || '');
     $('#nowtime').prop('checked', true);
     $('#eventTimeValue').val(moment().format('HH:mm'));
     $('#eventDateValue').val(moment().format('YYYY-MM-DD'));
@@ -30,18 +31,17 @@
     , insulin: $('#insulinGiven').val()
     , preBolus: parseInt($('#preBolus').val())
     , notes: $('#notes').val()
-    , units: browserSettings.units
+    , units: client.settings.units
     };
 
     if ($('#othertime').is(':checked')) {
-      data.eventTime = Nightscout.utils.mergeInputTime($('#eventTimeValue').val(), $('#eventDateValue').val());
+      data.eventTime = client.utils.mergeInputTime($('#eventTimeValue').val(), $('#eventDateValue').val());
     }
 
     return data;
   }
 
   function treatmentSubmit(event) {
-
     var data = prepareData();
 
     confirmPost(data);
@@ -84,14 +84,14 @@
       xhr.setRequestHeader('Content-Type', 'application/json; charset=UTF-8');
       xhr.send(dataJson);
 
-      browserStorage.set('enteredBy', data.enteredBy);
+      storage.set('enteredBy', data.enteredBy);
 
-      closeDrawer('#treatmentDrawer');
+      client.browserUtils.closeDrawer('#treatmentDrawer');
     }
   }
 
   $('#treatmentDrawerToggle').click(function (event) {
-    toggleDrawer('#treatmentDrawer', initTreatmentDrawer);
+    client.browserUtils.toggleDrawer('#treatmentDrawer', initTreatmentDrawer);
     event.preventDefault();
   });
 
@@ -107,26 +107,32 @@
     event.preventDefault();
   });
 
+  var eventTime = $('#eventTimeValue');
+  var eventDate = $('#eventDateValue');
+
   $('.eventinput').focus(function (event) {
     $('#othertime').prop('checked', true);
-    var moment = Nightscout.utils.mergeInputTime($('#eventTimeValue').val(), $('#eventDateValue').val());
+    var moment = client.utils.mergeInputTime(eventTime.val(), eventDate.val());
     $(this).attr('oldminutes', moment.minutes());
     $(this).attr('oldhours', moment.hours());
     event.preventDefault();
   })
   .change(function (event) {
     $('#othertime').prop('checked', true);
-    var moment = Nightscout.utils.mergeInputTime($('#eventTimeValue').val(), $('#eventDateValue').val());
+    var moment = client.utils.mergeInputTime(eventTime.val(), eventDate.val());
     if ($(this).attr('oldminutes') === ' 59' && moment.minutes() === 0) {
       moment.add(1, 'hours');
     }
     if ($(this).attr('oldminutes') === '0' && moment.minutes() === 59) {
       moment.add(-1, 'hours');
     }
-    $('#eventTimeValue').val(moment.format('HH:mm'));
-    $('#eventDateValue').val(moment.format('YYYY-MM-DD'));
+    eventTime.val(moment.format('HH:mm'));
+    eventDate.val(moment.format('YYYY-MM-DD'));
     $(this).attr('oldminutes', moment.minutes());
     $(this).attr('oldhours', moment.hours());
     event.preventDefault();
   });
-})();
+}
+
+module.exports = init;
+
