+++ lib/client/index.js
@@ -46,7 +46,8 @@ client.init = function init(serverSettings, plugins) {
 
   client.now = Date.now();
   client.data = [];
-  client.settings = require('./browser-settings')(serverSettings, plugins, $);
+  client.browserUtils = require('./browser-utils')($);
+  client.settings = require('./browser-settings')(client, plugins, serverSettings, $);
   client.utils = require('../utils')(client.settings);
 
   client.sbx = sandbox.clientInit(client.settings, client.now);
@@ -68,6 +69,8 @@ client.init = function init(serverSettings, plugins) {
   client.dataUpdate = dataUpdate;
 
   client.renderer = require('./renderer')(client, d3, $);
+  client.careportal = require('./careportal')(client, $);
+
   var timeAgo = client.utils.timeAgo;
 
   var container = $('.container')
@@ -423,10 +426,10 @@ client.init = function init(serverSettings, plugins) {
 
   function playAlarm(audio) {
     // ?mute=true disables alarms to testers.
-    if (querystring.mute !== 'true') {
+    if (client.browserUtils.queryParms().mute !== 'true') {
       audio.play();
     } else {
-      showNotification('Alarm was muted (?mute=true)');
+      client.browserUtils.showNotification('Alarm was muted (?mute=true)');
     }
   }
 
@@ -440,7 +443,7 @@ client.init = function init(serverSettings, plugins) {
       $(this).removeClass('playing');
     });
 
-    closeNotification();
+    client.browserUtils.closeNotification();
     container.removeClass('alarming');
 
     updateTitle();
