+++ static/index.html
@@ -263,8 +263,6 @@
         <script src="/bower_components/d3/d3.min.js"></script>
         <script src="/bower_components/jQuery-Storage-API/jquery.storageapi.min.js"></script>
         <script src="/bower_components/tipsy-jmalonzo/src/javascripts/jquery.tipsy.js"></script>
-        <script src="/js/client.js?v=0.7.0"></script>
-        <script src="/js/ui-utils.js?v=0.7.0"></script>
-        <script src="/js/treatment.js"></script>
+        <script src="/js/client.js"></script>
     </body>
 </html>
