--- static/js/ui-utils.js
@@ -1,113 +1,0 @@
-'use strict';
-
-var openDraw = null;
-
-function getQueryParms() {
-  var params = {};
-  if (location.search) {
-    location.search.substr(1).split('&').forEach(function(item) {
-      params[item.split('=')[0]] = item.split('=')[1].replace(/[_\+]/g, ' ');
-    });
-  }
-  return params;
-}
-
-function isTouch() {
-  try { document.createEvent('TouchEvent'); return true; }
-  catch (e) { return false; }
-}
-
-function closeDrawer(id, callback) {
-  openDraw = null;
-  $('html, body').animate({ scrollTop: 0 });
-  $(id).animate({right: '-300px'}, 300, function () {
-    $(id).css('display', 'none');
-    if (callback) { callback(); }
-  });
-}
-
-function toggleDrawer(id, openCallback, closeCallback) {
-
-  function openDrawer(id, callback) {
-    function closeOpenDraw(callback) {
-      if (openDraw) {
-        closeDrawer(openDraw, callback);
-      } else {
-        callback();
-      }
-    }
-
-    closeOpenDraw(function () {
-      openDraw = id;
-      $(id).css('display', 'block').animate({right: '0'}, 300, function () {
-        if (callback) { callback(); }
-      });
-    });
-
-  }
-
-  if (openDraw === id) {
-    closeDrawer(id, closeCallback);
-  } else {
-    openDrawer(id, openCallback);
-  }
-
-}
-
-function closeNotification() {
-  var notify = $('#notification');
-  notify.hide();
-  notify.find('span').html('');
-}
-
-function showNotification(note, type)  {
-  var notify = $('#notification');
-  notify.hide();
-
-  // Notification types: 'info', 'warn', 'success', 'urgent'.
-  // - default: 'urgent'
-  notify.removeClass('info warn urgent');
-  notify.addClass(type ? type : 'urgent');
-
-  notify.find('span').html(note);
-  notify.css('left', 'calc(50% - ' + (notify.width() / 2) + 'px)');
-  notify.show();
-}
-
-
-var querystring = getQueryParms();
-
-$('#drawerToggle').click(function(event) {
-  toggleDrawer('#drawer');
-  event.preventDefault();
-});
-
-$('#notification').click(function(event) {
-  closeNotification();
-  event.preventDefault();
-});
-
-$(function() {
-  // Tooltips can remain in the way on touch screens.
-  var notTouchScreen = (!isTouch());
-  if (notTouchScreen) {
-    $('.tip').tipsy();
-  } else {
-    // Drawer info tips should be displayed on touchscreens.
-    $('#drawer').find('.tip').tipsy();
-  }
-  $.fn.tipsy.defaults = {
-    fade: true,
-    gravity: 'n',
-    opacity: 0.75
-  };
-
-  if (querystring.notify) {
-    showNotification(querystring.notify, querystring.notifytype);
-  }
-
-  if (querystring.drawer) {
-    openDrawer('#drawer');
-  }
-});
-
